--- src/pages/topic-details.tsx
@@ -1,15 +1,0 @@
-import Layout from "../hoc/layout";
-
-import TopicDetails from "../components/ComponentPages/TopicDetails";
-
-const TopicDetailsPage = () => {
-  return (
-    <>
-      <Layout>
-        <TopicDetails />
-      </Layout>
-    </>
-  );
-};
-
-export default TopicDetailsPage;
